feat(replays): Show Errors/Exceptions in the list of replay breadcrumbs
