+++ static/app/components/events/interfaces/breadcrumbs/index.tsx
@@ -15,7 +15,7 @@ import {
   Crumb,
   RawCrumb,
 } from 'sentry/types/breadcrumbs';
-import {EntryType, Event} from 'sentry/types/event';
+import {Event} from 'sentry/types/event';
 import {defined} from 'sentry/utils';
 
 import SearchBarAction from '../searchBarAction';
@@ -24,7 +24,7 @@ import SearchBarActionFilter from '../searchBarAction/searchBarActionFilter';
 import Level from './breadcrumb/level';
 import Type from './breadcrumb/type';
 import Breadcrumbs from './breadcrumbs';
-import {transformCrumbs} from './utils';
+import {getVirtualCrumb, transformCrumbs} from './utils';
 
 type FilterOptions = React.ComponentProps<typeof SearchBarActionFilter>['options'];
 
@@ -84,13 +84,13 @@ function BreadcrumbsContainer({
 
   useEffect(() => {
     loadBreadcrumbs();
-  }, []);
+  }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
   function loadBreadcrumbs() {
     let crumbs = data.values;
 
     // Add the (virtual) breadcrumb based on the error or message event if possible.
-    const virtualCrumb = getVirtualCrumb();
+    const virtualCrumb = getVirtualCrumb(event);
 
     if (virtualCrumb) {
       crumbs = [...crumbs, virtualCrumb];
@@ -241,51 +241,6 @@ function BreadcrumbsContainer({
     }
 
     return breadcrumbs;
-  }
-
-  function moduleToCategory(module?: string | null) {
-    if (!module) {
-      return undefined;
-    }
-    const match = module.match(/^.*\/(.*?)(:\d+)/);
-    if (!match) {
-      return module.split(/./)[0];
-    }
-    return match[1];
-  }
-
-  function getVirtualCrumb(): RawCrumb | undefined {
-    const exception = event.entries.find(entry => entry.type === EntryType.EXCEPTION);
-
-    if (!exception && !event.message) {
-      return undefined;
-    }
-
-    const timestamp = event.dateCreated;
-
-    if (exception) {
-      const {type, value, module: mdl} = exception.data.values[0];
-      return {
-        type: BreadcrumbType.ERROR,
-        level: BreadcrumbLevelType.ERROR,
-        category: moduleToCategory(mdl) || 'exception',
-        data: {
-          type,
-          value,
-        },
-        timestamp,
-      };
-    }
-
-    const levelTag = (event.tags || []).find(tag => tag.key === 'level');
-
-    return {
-      type: BreadcrumbType.INFO,
-      level: (levelTag?.value as BreadcrumbLevelType) || BreadcrumbLevelType.UNDEFINED,
-      category: 'message',
-      message: event.message,
-      timestamp,
-    };
   }
 
   function handleSearch(value: string) {
